session-2/requirements.txt
matplotlib
numpy
Pillow
torch
ray
scikit-learn
scikit-image
torchvision
pandas
ray[tune]